Move index related files to index dir
The files are:
- invertedindex.cpp(h)
- seriesparser.cpp(h)
- stringpool.cpp(h)
