+++ libakumuli/index/invertedindex.cpp
