+++ libakumuli/index/invertedindex.h
