+++ libakumuli/index/seriesparser.cpp
