+++ libakumuli/index/stringpool.cpp
