+++ libakumuli/index/stringpool.h
