+++ unittests/CMakeLists.txt
@@ -177,8 +177,8 @@ add_executable(
     ../libakumuli/util.cpp
     ../libakumuli/datetime.cpp
     ../libakumuli/log_iface.cpp
-    ../libakumuli/seriesparser.cpp
-    ../libakumuli/stringpool.cpp
+    ../libakumuli/index/seriesparser.cpp
+    ../libakumuli/index/stringpool.cpp
     ../libakumuli/crc32c.cpp
     ../libakumuli/status_util.cpp
     ../libakumuli/storage_engine/blockstore.cpp
@@ -224,8 +224,8 @@ add_test(storage test_storage)
 add_executable(
     test_seriesparser
     test_parser.cpp
-    ../libakumuli/seriesparser.cpp
-    ../libakumuli/stringpool.cpp
+    ../libakumuli/index/seriesparser.cpp
+    ../libakumuli/index/stringpool.cpp
     ../libakumuli/util.cpp
     ../libakumuli/log_iface.cpp
     ../libakumuli/datetime.cpp
@@ -249,8 +249,8 @@ add_executable(
     ../libakumuli/saxencoder.cpp
     ../libakumuli/anomalydetector.cpp
     ../libakumuli/hashfnfamily.cpp
-    ../libakumuli/seriesparser.cpp
-    ../libakumuli/stringpool.cpp
+    ../libakumuli/index/seriesparser.cpp
+    ../libakumuli/index/stringpool.cpp
     ../libakumuli/util.cpp
     ../libakumuli/log_iface.cpp
     ../libakumuli/datetime.cpp
@@ -372,8 +372,8 @@ add_executable(
     ../libakumuli/status_util.cpp
     ../libakumuli/log_iface.cpp
     ../libakumuli/crc32c.cpp
-    ../libakumuli/seriesparser.cpp
-    ../libakumuli/stringpool.cpp
+    ../libakumuli/index/seriesparser.cpp
+    ../libakumuli/index/stringpool.cpp
     ../libakumuli/metadatastorage.cpp
 )
