+++ unittests/test_parser.cpp
@@ -4,7 +4,7 @@
 #define BOOST_TEST_MODULE Main
 #include <boost/test/unit_test.hpp>
 
-#include "seriesparser.h"
+#include "index/seriesparser.h"
 #include "queryprocessor_framework.h"
 #include "datetime.h"
 #include <tuple>
